Move SoftiMax reader to io/maxiv.py
