+++ orangecontrib/spectroscopy/io/__init__.py
@@ -21,6 +21,6 @@ from .perkinelmer import PerkinElmerReader
 
 # Facility-specific readers
 from .diamond import NXS_STXM_Diamond_I08
-from .maxiv import HDRReader_STXM
+from .maxiv import HDRReader_STXM, HDF5Reader_SoftiMAX
 from .soleil import SelectColumnReader, HDF5Reader_HERMES, HDF5Reader_ROCK
 from .cls import HDF5Reader_SGM
